+++ core/config-manager.js
@@ -10,7 +10,7 @@ const pathManager = require('./path-manager')
  * - Create new configurations
  * - Update existing configurations
  *
- * @version 0.2.0
+ * @version 0.2.1
  */
 class ConfigManager {
   /**